5100: return chart, applied figma styles
